Create src/utils/protocols.ts
